requirements.txt
git+git://github.com/ejhumphrey/claudio.git
pescador
progressbar2
pyyaml
